requirements.txt
requests
python-dotenv
pandas
streamlit
py-clob-client
scikit-learn
numpy
newsapi-python
